+++ src/pages/configuration.py
@@ -1,10 +1,11 @@
+import os
 from enum import Enum
 import streamlit as st
-from schema import Triple, Endpoint, Graph
+from schema import Triple, Endpoint, Graph, Prefix
 from lib.sparql_base import delete
 from lib.sparql_queries import count_graph_triples
 from lib.utils import readable_number, ensure_uri
-from lib.configuration import save_config, get_config_toml, CONFIG_PATH
+from lib.configuration import save_config, unload_config, CONFIG_PATH
 import lib.state as state
 from components.init import init
 from components.menu import menu
@@ -12,6 +13,7 @@ from components.dialog_confirmation import dialog_confirmation
 from components.dialog_download import dialog_download_graph, dialog_dump_endpoint
 from components.dialog_config_endpoint import dialog_config_endpoint
 from components.dialog_config_graph import dialog_config_graph
+from components.dialog_config_prefix import dialog_config_prefix
 
 
 def __delete_endpoint(endpoint: Endpoint) -> None:
@@ -21,7 +23,7 @@ def __delete_endpoint(endpoint: Endpoint) -> None:
     state.delete_endpoint(endpoint)
 
     # Save config if local
-    if state.get_configuration() == 'local':
+    if os.getenv('ENV') != 'streamlit':
         save_config()
 
     # Close confirmation window and reload the page with new state
@@ -46,6 +48,21 @@ def __prune_graph(graph: Graph) -> None:
 
     state.clear_graphs()
     state.set_toast('Graph deleted', icon=':material/delete:')
+    st.rerun()
+
+
+def __delete_prefix(prefix: Prefix) -> None:
+    """Delete the specified endpoint"""
+
+    # Delete endpoint from state
+    state.delete_prefix(prefix)
+
+    # Save config if local
+    if os.getenv('ENV') != 'streamlit':
+        save_config()
+
+    # Close confirmation window and reload the page with new state
+    state.set_toast('Prefix removed', icon=':material/delete:')
     st.rerun()
 
 
@@ -150,6 +167,44 @@ def __graph_list() -> None:
         if col5.button('', key=f"config-graph-download-csv-{i}", icon=":material/download:", type='tertiary'):
             dialog_download_graph(graph)
 
+
+def __show_prefix_list() -> None:
+    # Here, because it is only for here and for display information
+    # We took the liberty of handling the session_state directly
+    st.session_state['display_prefix_list'] = True
+
+
+def __hide_prefix_list() -> None:
+    # Here, because it is only for here and for display information
+    # We took the liberty of handling the session_state directly
+    st.session_state['display_prefix_list'] = False
+
+
+def __prefix_list() -> None:
+
+    # If graph list is deactivated, display nothing
+    # Here, because it is only for here and for display information
+    # We took the liberty of handling the session_state directly
+    if 'display_prefix_list'not in st.session_state or not st.session_state['display_prefix_list']:
+        return
+    
+    # Fetch from state
+    all_prefixes = state.get_prefixes()
+
+    # Display all graphs in sessions
+    for i, prefix in enumerate(all_prefixes):
+        st.text('')
+        col1, col2, col3, col4 = st.columns([2, 7, 1, 1], vertical_alignment='center')
+        col1.markdown(prefix.short)
+        col2.markdown(prefix.url)
+
+        # Button to cleanse a graph
+        if col3.button('', key=f"config-prefix-delete-{i}", icon=":material/delete:", help='This will delete this prefix from your configuration.', type='tertiary'):
+            dialog_confirmation(f'You are about to delete this prefix from your configuration.', __delete_prefix, prefix=prefix)
+
+
+
+
 ##### The page #####
 
 init()
@@ -165,10 +220,10 @@ col1, col2 = st.columns([6, 2], vertical_alignment='bottom')
 col1.title("Endpoint configuration")
 col2.download_button(
     label='Download', 
-    data=get_config_toml(), 
-    disabled=state.get_configuration() != 'uploaded',
+    data=unload_config(), 
+    disabled=os.getenv('ENV') != 'streamlit',
     file_name=CONFIG_PATH,
-    help='Download this configuration as a toml file. If Logre is running locally and have a configuration, it will be updated automatically.',
+    help='Download this configuration as a toml file. If Logre is running locally, it will be updated automatically.',
     icon=':material/download:'
 )
 
@@ -179,8 +234,6 @@ st.divider()
 
 col1, col2, col3, col4 = st.columns([4, 1, 1, 2], vertical_alignment='bottom')
 col1.markdown('### Endpoints')
-
-st.text("")
 
 col2.button('Show', on_click=__show_endpoint_list, icon=':material/visibility:', key='config-btn-show-endpoints', type='tertiary')
 col3.button('Hide', on_click=__hide_endpoint_list, icon=':material/visibility_off:', key='config-btn-hide-endpoints', type='tertiary')
@@ -209,3 +262,19 @@ if endpoint:
     __graph_list()
 
     st.divider()
+
+
+### PREFIXES ###
+
+
+# Title and boxes for graph actions (show/hide graph list)
+col1, col_dump, col2, col3, col4 = st.columns([2, 2, 1, 1, 2], vertical_alignment='bottom')
+col1.markdown('### Prefixes')
+
+col2.button('Show', on_click=__show_prefix_list, icon=':material/visibility:', key='config-btn-show-prefixes', type='tertiary')
+col3.button('Hide', on_click=__hide_prefix_list, icon=':material/visibility_off:', key='config-btn-hide-prefixes', type='tertiary')
+col4.button('Add new', on_click=dialog_config_prefix, icon=':material/add:', key='config-btn-add-prefix')
+
+__prefix_list()
+
+st.divider()